src/requirements.txt
aioredis==2.0.0
aiosmtplib==1.1.6
alembic==1.7.3
anyio==3.4.0
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
databases==0.5.3
Deprecated==1.2.13
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fakeredis==1.7.0
fastapi==0.70.0
fastapi-mail==1.0.2
fastapi-users==8.1.1
fastapi-users-db-sqlalchemy==1.0.0
greenlet==1.1.2
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.3
importlib-resources==5.2.2
iniconfig==1.1.1
Jinja2==3.0.3
makefun==1.12.1
Mako==1.1.5
MarkupSafe==2.0.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.11.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
PyJWT==2.2.0
pyparsing==3.0.6
pytest==6.2.5
python-decouple==3.5
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
redis==4.0.2
requests==2.26.0
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.25
SQLAlchemy-ImageAttach==1.1.0
starlette==0.16.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
Wand==0.6.7
wrapt==1.13.3
zipp==3.6.0
